+++ src/main/java/com/clickhouse/kafka/connect/sink/state/provider/KeeperStateProvider.java
@@ -83,7 +83,7 @@ public class KeeperStateProvider implements StateProvider {
                      .format(ClickHouseFormat.RowBinaryWithNamesAndTypes)
                      .query(selectStr)
                      .executeAndWait()) {
-            LOGGER.debug("return size: " + response.getSummary().getReadRows());
+            LOGGER.debug("return size: {}", response.getSummary().getReadRows());
             if ( response.getSummary().getReadRows() == 0) {
                 LOGGER.info(String.format("read state record: topic %s partition %s with NONE state", topic, partition));
                 return new StateRecord(topic, partition, 0, 0, State.NONE);